+++ crates/environ/src/vmoffsets.rs
@@ -29,7 +29,7 @@ use std::convert::TryFrom;
 /// Sentinel value indicating that wasm has been interrupted.
 // Note that this has a bit of an odd definition. See the `insert_stack_check`
 // function in `cranelift/codegen/src/isa/x86/abi.rs` for more information
-pub const INTERRUPTED: i32 = (u32::max_value() - 32 * 1024) as i32;
+pub const INTERRUPTED: usize = usize::max_value() - 32 * 1024;
 
 #[cfg(target_pointer_width = "32")]
 fn cast_to_u32(sz: usize) -> u32 {
